refactor: add stateful parsers for composer.lock and package-lock.json files
